requirements.txt
streamlit==1.35.0
easyocr==1.7.1
googletrans==4.0.0-rc1
Pillow==10.2.0
numpy==1.24.4
torch==2.0.1
opencv-python-headless==4.6.0.66